move DeveloperCardComponent to Developers namespace
