--- app/components/developer_card_component.rb
@@ -1,22 +1,0 @@
-class DeveloperCardComponent < ApplicationComponent
-  with_collection_parameter :developer
-
-  attr_reader :developer
-
-  def initialize(developer:, featured: false)
-    @developer = developer
-    @featured = featured
-  end
-
-  def hero
-    @developer.hero
-  end
-
-  def bio
-    @developer.bio
-  end
-
-  def featured?
-    !!@featured
-  end
-end
