+++ app/components/developers/card_component.html.erb
