+++ app/views/developers/index.html.erb
@@ -41,7 +41,7 @@
 
     <!-- Paginated list of developers -->
     <div class="col-span-12 md:col-span-9">
-      <%= render DeveloperCardComponent.with_collection(@query.featured_records, featured: true) %>
+      <%= render Developers::CardComponent.with_collection(@query.featured_records, featured: true) %>
 
       <%= render PagyPaginatorComponent.new(id: "developers", pagy: @query.pagy, url_array: [:developers], container_classes: "divide-y divide-gray-200", options: @query.filters) do |component| %>
         <% component.loading_icon do %>
@@ -49,7 +49,7 @@
             <%= render LoadingComponent.new %>
           </div>
         <% end %>
-        <%= render DeveloperCardComponent.with_collection(@query.records) %>
+        <%= render Developers::CardComponent.with_collection(@query.records) %>
       <% end %>
     </div>
   </div>
