+++ app/views/home/show.html.erb
@@ -30,7 +30,7 @@
   </div>
 
   <div class="shadow overflow-hidden sm:rounded-md divide-y divide-gray-200">
-    <%= render DeveloperCardComponent.with_collection(@developers) %>
+    <%= render Developers::CardComponent.with_collection(@developers) %>
   </div>
 
   <div class="mt-8 mb-16 text-right mr-2 lg:mr-0">
